+++ test/core/test_timecoord.py
@@ -48,21 +48,70 @@ class AddTimeCoordsTest(unittest.TestCase):
 
     def test_from_time_in_days_since_1970(self):
         self.assertEqual('2017-06-07T12:00:00.000000000',
-                         str(from_time_in_days_since_1970(to_time_in_days_since_1970('201706071200'))))
+                         str(from_time_in_days_since_1970(
+                             to_time_in_days_since_1970('201706071200'))))
         self.assertEqual('2017-06-08T12:00:00.000000000',
-                         str(from_time_in_days_since_1970(to_time_in_days_since_1970('201706081200'))))
+                         str(from_time_in_days_since_1970(
+                             to_time_in_days_since_1970('201706081200'))))
         self.assertEqual('2018-06-08T12:00:00.000000000',
-                         str(from_time_in_days_since_1970(to_time_in_days_since_1970('2018-06-08 12:00'))))
+                         str(from_time_in_days_since_1970(
+                             to_time_in_days_since_1970('2018-06-08 12:00'))))
         self.assertEqual('2018-06-08T12:00:00.000000000',
-                         str(from_time_in_days_since_1970(to_time_in_days_since_1970('2018-06-08T12:00'))))
+                         str(from_time_in_days_since_1970(
+                             to_time_in_days_since_1970('2018-06-08T12:00'))))
         self.assertEqual('2019-10-04T10:13:48.538000000',
-                         str(from_time_in_days_since_1970(to_time_in_days_since_1970('04-OCT-2019 10:13:48.538184'))))
+                         str(from_time_in_days_since_1970(
+                             to_time_in_days_since_1970('04-OCT-2019 10:13:48.538184'))))
 
 
 class GetTimeRangeTest(unittest.TestCase):
 
     def test_get_time_range_from_data(self):
         cube = new_cube(drop_bounds=True)
+        time_range = get_time_range_from_data(cube)
+        self.assertIsNotNone(time_range)
+        self.assertEqual('2010-01-01T00:00:00', pd.Timestamp(time_range[0]).isoformat())
+        self.assertEqual('2010-01-06T00:00:00', pd.Timestamp(time_range[1]).isoformat())
+
+    def test_get_time_range_from_data_cftime(self):
+        cube = new_cube(drop_bounds=True,
+                        use_cftime=True,
+                        time_dtype=None)
+        time_range = get_time_range_from_data(cube)
+        self.assertIsNotNone(time_range)
+        self.assertEqual('2010-01-01T00:00:00', pd.Timestamp(time_range[0]).isoformat())
+        self.assertEqual('2010-01-06T00:00:00', pd.Timestamp(time_range[1]).isoformat())
+
+    def test_get_time_range_from_data_time_named_t(self):
+        cube = new_cube(drop_bounds=True, time_name='t')
+        time_range = get_time_range_from_data(cube)
+        self.assertIsNotNone(time_range)
+        self.assertEqual('2010-01-01T00:00:00', pd.Timestamp(time_range[0]).isoformat())
+        self.assertEqual('2010-01-06T00:00:00', pd.Timestamp(time_range[1]).isoformat())
+
+    def test_get_time_range_from_data_additional_t_variable(self):
+        import xarray as xr
+        start_time_data = pd.date_range(start='2010-01-03T12:00:00',
+                                        periods=5,
+                                        freq='5D').values.astype(dtype='datetime64[s]')
+        start_time = xr.DataArray(start_time_data, dims='time')
+        end_time_data = pd.date_range(start='2010-01-07T12:00:00',
+                                      periods=5,
+                                      freq='5D').values.astype(dtype='datetime64[s]')
+        end_time = xr.DataArray(end_time_data, dims='time')
+        cube = new_cube(drop_bounds=True,
+                        time_start='2010-01-05T12:00:00',
+                        time_freq='5D',
+                        variables=dict(start_time=start_time, end_time=end_time))
+        time_range = get_time_range_from_data(cube)
+        self.assertIsNotNone(time_range)
+        self.assertEqual('2010-01-03T12:00:00', pd.Timestamp(time_range[0]).isoformat())
+        self.assertEqual('2010-01-27T12:00:00', pd.Timestamp(time_range[1]).isoformat())
+
+    def test_get_time_range_from_data_start_and_end_time_arrays(self):
+        cube = new_cube(drop_bounds=True,
+                        use_cftime=True,
+                        time_dtype=None)
         time_range = get_time_range_from_data(cube)
         self.assertIsNotNone(time_range)
         self.assertEqual('2010-01-01T00:00:00', pd.Timestamp(time_range[0]).isoformat())
@@ -106,10 +155,14 @@ class TimestampToIsoStringTest(unittest.TestCase):
 
     def test_it_with_h_res(self):
         self.assertEqual("2018-09-05T00:00:00Z",
-                         timestamp_to_iso_string(np.datetime64("2018-09-05"), freq="H"))
+                         timestamp_to_iso_string(np.datetime64("2018-09-05"),
+                                                 freq="H"))
         self.assertEqual("2018-09-05T11:00:00Z",
-                         timestamp_to_iso_string(np.datetime64("2018-09-05 10:35:42"), freq="H"))
+                         timestamp_to_iso_string(np.datetime64("2018-09-05 10:35:42"),
+                                                 freq="H"))
         self.assertEqual("2018-09-05T11:00:00Z",
-                         timestamp_to_iso_string(np.datetime64("2018-09-05 10:35:42.164"), freq="H"))
+                         timestamp_to_iso_string(np.datetime64("2018-09-05 10:35:42.164"),
+                                                 freq="H"))
         self.assertEqual("2019-10-04T10:00:00Z",
-                         timestamp_to_iso_string(pd.to_datetime("04-OCT-2019 10:13:48.538184"), freq="H"))
+                         timestamp_to_iso_string(pd.to_datetime("04-OCT-2019 10:13:48.538184"),
+                                                 freq="H"))